HomeWork2/requirements.txt
pytesseract
pillow


